pyproject.toml
[project]
name = "hope-smart-import"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django-admin-extra-buttons>=1.5.8",
    "hope-flex-fields",
    "openpyxl>=3.1.5",
]

[project.optional-dependencies]
docs = [
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.5.36",
    "mkdocs-awesome-pages-plugin>=2.9.3",
    "mkdocstrings-python",
]


[tool.uv]
dev-dependencies = [
    "django-pytest>=0.2.0",
    "django-webtest>=1.9.12",
    "pytest>=8.3.3",
    "pytest-cov>=5.0.0",
    "pytest-echo>=1.7.3",
    "factory-boy>=3.3.1",
    "pytest-django>=4.9.0",
    "pdbpp>=0.10.3",
    "ipython>=8.27.0",
    "isort>=5.13.2",
    "flake8>=7.1.1",
    "black>=24.8.0",
    "django>=4,<5",
    "mypy>=1.11.2",
]

[tool.uv.sources]
hope-flex-fields = { path = "../hope-flex-fields" }




[tool.isort]
profile = "black"

[tool.black]
line-length = 120
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | venv
  | _build
  | buck-out
  | build
  | dist
  | migrations
  | snapshots
)/
'''
